% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_earthdata_token}
\alias{rust_earthdata_token}
\title{Exchange Earthdata username/password for a bearer token.}
\usage{
rust_earthdata_token(username, password)
}
\description{
Calls the NASA Earthdata Login token API. Returns the access token string.
}
