% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_hdf5_dataset}
\alias{rust_hdf5_dataset}
\title{Read a single dataset from an HDF5 file and return raw bytes + metadata.}
\usage{
rust_hdf5_dataset(url, dataset_path, username, password)
}
\description{
Read a single dataset from an HDF5 file and return raw bytes + metadata.
}
