% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_hdf5_groups}
\alias{rust_hdf5_groups}
\title{List available groups in an HDF5 file (for exploration).}
\usage{
rust_hdf5_groups(url, path, username, password)
}
\description{
List available groups in an HDF5 file (for exploration).
}
