% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_read_gedi}
\alias{rust_read_gedi}
\title{Read GEDI data from a remote HDF5 file with spatial subsetting.}
\usage{
rust_read_gedi(
  url,
  product,
  xmin,
  ymin,
  xmax,
  ymax,
  columns,
  beams,
  username,
  password
)
}
\description{
Read GEDI data from a remote HDF5 file with spatial subsetting.
}
