% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_read_icesat2}
\alias{rust_read_icesat2}
\title{Read ICESat-2 data from a remote HDF5 file with spatial subsetting.}
\usage{
rust_read_icesat2(
  url,
  product,
  xmin,
  ymin,
  xmax,
  ymax,
  columns,
  tracks,
  username,
  password
)
}
\description{
Read ICESat-2 data from a remote HDF5 file with spatial subsetting.
}
