% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_read_icesat2_multi}
\alias{rust_read_icesat2_multi}
\title{Read ICESat-2 data from multiple remote HDF5 files concurrently.}
\usage{
rust_read_icesat2_multi(
  urls,
  product,
  xmin,
  ymin,
  xmax,
  ymax,
  columns,
  tracks,
  username,
  password,
  pool_columns,
  transposed_columns
)
}
\description{
Read ICESat-2 data from multiple remote HDF5 files concurrently.
}
