% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sl_bbox}
\alias{sl_bbox}
\title{Create a bounding box for spatial queries}
\usage{
sl_bbox(xmin, ymin, xmax, ymax)
}
\arguments{
\item{xmin}{Minimum longitude (western boundary).}

\item{ymin}{Minimum latitude (southern boundary).}

\item{xmax}{Maximum longitude (eastern boundary).}

\item{ymax}{Maximum latitude (northern boundary).}
}
\value{
A named double vector of class \code{sl_bbox}.
}
\description{
Create a bounding box for spatial queries
}
