% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacelaser-package.R
\docType{package}
\name{spacelaser-package}
\alias{spacelaser}
\alias{spacelaser-package}
\title{spacelaser: Cloud-Optimized Partial Reading of GEDI and ICESat-2 HDF5 Data}
\description{
Read GEDI and ICESat-2 satellite lidar data directly from NASA Earthdata cloud storage using HTTP range requests. Only fetches the bytes needed for the requested spatial subset, avoiding multi-gigabyte file downloads. Powered by a pure-Rust h5coro-style HDF5 partial reader.
}
\author{
\strong{Maintainer}: Hugh Graham \email{hugh@belian.earth}

Other contributors:
\itemize{
  \item belian.earth [copyright holder]
}

}
\keyword{internal}
